backend/HrayfiConnect_Mobile/artisan-platform/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
motor>=3.3.0
pymongo>=4.5.0
bcrypt>=4.0.0
python-jose>=3.3.0
passlib>=1.7.4
cryptography>=41.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
python-multipart>=0.0.6
PyJWT>=2.8.0
cloudinary>=1.38.0
websockets>=10.0